rename mcp_tools_mange_service.py to mcp_tools_manage_service.py
